dev-requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --extra=dev --extra=doc --extra=test --output-file=dev-requirements.txt pyproject.toml
#
aiohttp==3.7.4.post0
    # via
    #   discord-py-slash-command
    #   discord.py
    #   sxm
alabaster==0.7.12
    # via sphinx
anyio==3.2.1
    # via httpcore
appdirs==1.4.4
    # via black
astroid==2.6.2
    # via pylint
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via
    #   aiohttp
    #   flake8-bugbear
    #   pytest
babel==2.9.1
    # via sphinx
backcall==0.2.0
    # via ipython
backports.entry-points-selectable==1.1.0
    # via virtualenv
bandit==1.7.0
    # via sxm_discord (pyproject.toml)
beautifulsoup4==4.9.3
    # via furo
black==21.7b0
    # via sxm_discord (pyproject.toml)
certifi==2021.5.30
    # via
    #   httpx
    #   requests
cffi==1.14.6
    # via pynacl
chardet==4.0.0
    # via aiohttp
charset-normalizer==2.0.3
    # via requests
click==7.1.2
    # via
    #   black
    #   pip-tools
    #   typer
colorama==0.4.4
    # via rich
coloredlogs==15.0.1
    # via sxm-player
commonmark==0.9.1
    # via rich
coverage==5.5
    # via
    #   pytest-cov
    #   sxm_discord (pyproject.toml)
decorator==4.4.2
    # via
    #   ipython
    #   networkx
discord-py-slash-command==2.3.2
    # via sxm_discord (pyproject.toml)
discord.py[voice]==1.7.3
    # via
    #   discord-py-slash-command
    #   discord.py-stubs
    #   sxm_discord (pyproject.toml)
discord.py-stubs==1.7.3
    # via sxm_discord (pyproject.toml)
distlib==0.3.2
    # via virtualenv
dnspython==1.16.0
    # via pymongo
docopt==0.6.2
    # via snooty-lextudio
docutils==0.17.1
    # via
    #   flit
    #   rstcheck
    #   snooty-lextudio
    #   sphinx
fake-useragent==0.1.11
    # via sxm
fett==0.3.2
    # via snooty-lextudio
filelock==3.0.12
    # via
    #   tox
    #   virtualenv
flake8==3.9.2
    # via
    #   flake8-bugbear
    #   flake8-builtins
    #   flake8-commas
    #   flake8-comprehensions
    #   pyproject-flake8
    #   sxm_discord (pyproject.toml)
flake8-bugbear==21.4.3
    # via sxm_discord (pyproject.toml)
flake8-builtins==1.5.3
    # via sxm_discord (pyproject.toml)
flake8-commas==2.0.0
    # via sxm_discord (pyproject.toml)
flake8-comprehensions==3.5.0
    # via sxm_discord (pyproject.toml)
flit==3.2.0
    # via sxm_discord (pyproject.toml)
flit-core==3.2.0
    # via flit
furo==2021.7.5b38
    # via sxm_discord (pyproject.toml)
gitdb==4.0.7
    # via gitpython
gitpython==3.1.18
    # via bandit
greenlet==1.1.0
    # via sqlalchemy
h11==0.12.0
    # via httpcore
httpcore==0.13.6
    # via httpx
httpx==0.18.2
    # via
    #   sxm
    #   sxm-player
humanfriendly==9.2
    # via coloredlogs
humanize==3.10.0
    # via sxm_discord (pyproject.toml)
idna==3.2
    # via
    #   anyio
    #   requests
    #   rfc3986
    #   yarl
imagesize==1.2.0
    # via sphinx
iniconfig==1.1.1
    # via pytest
ipython==7.25.0
    # via sxm-player
ipython-genutils==0.2.0
    # via traitlets
isort==5.9.2
    # via
    #   pylint
    #   sxm_discord (pyproject.toml)
jedi==0.18.0
    # via ipython
jinja2==3.0.1
    # via sphinx
lazy-object-proxy==1.6.0
    # via astroid
make-it-sync==1.0.0
    # via sxm
markupsafe==2.0.1
    # via jinja2
matplotlib-inline==0.1.2
    # via ipython
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
mypy==0.910
    # via
    #   discord.py-stubs
    #   sqlalchemy-stubs
    #   sxm_discord (pyproject.toml)
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
networkx==2.5.1
    # via snooty-lextudio
packaging==21.0
    # via
    #   pytest
    #   pytest-sugar
    #   sphinx
    #   tox
parso==0.8.2
    # via jedi
pathspec==0.8.1
    # via black
pbr==5.6.0
    # via stevedore
pep517==0.10.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.2.0
    # via sxm_discord (pyproject.toml)
platformdirs==2.0.2
    # via virtualenv
pluggy==0.13.1
    # via
    #   pytest
    #   tox
pprintpp==0.4.0
    # via pytest-clarity
prompt-toolkit==3.0.19
    # via ipython
psutil==5.8.0
    # via sxm-player
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via
    #   pytest
    #   tox
pycodestyle==2.7.0
    # via flake8
pycparser==2.20
    # via cffi
pydantic==1.8.2
    # via
    #   sxm
    #   sxm-player
pyflakes==2.3.1
    # via flake8
pygments==2.9.0
    # via
    #   ipython
    #   rich
    #   sphinx
pylint==2.9.3
    # via sxm_discord (pyproject.toml)
pymongo[srv,tls]==3.11.4
    # via snooty-lextudio
pynacl==1.4.0
    # via discord.py
pyparsing==2.4.7
    # via packaging
pyproject-flake8==0.0.1a2
    # via sxm_discord (pyproject.toml)
pytest==6.2.4
    # via
    #   pytest-clarity
    #   pytest-cov
    #   pytest-socket
    #   pytest-sugar
    #   sxm_discord (pyproject.toml)
pytest-clarity==1.0.1
    # via sxm_discord (pyproject.toml)
pytest-cov==2.12.1
    # via sxm_discord (pyproject.toml)
pytest-socket==0.4.0
    # via sxm_discord (pyproject.toml)
pytest-sugar==0.9.4
    # via sxm_discord (pyproject.toml)
python-dotenv==0.18.0
    # via sxm
python-jsonrpc-server==0.3.4
    # via snooty-lextudio
pytz==2021.1
    # via babel
pyyaml==5.4.1
    # via
    #   bandit
    #   snooty-lextudio
    #   sxm-player
regex==2021.7.6
    # via black
requests==2.26.0
    # via
    #   flit
    #   snooty-lextudio
    #   sphinx
rfc3986[idna2008]==1.5.0
    # via httpx
rich==10.6.0
    # via pytest-clarity
rstcheck==3.3.1
    # via sxm_discord (pyproject.toml)
six==1.16.0
    # via
    #   bandit
    #   pynacl
    #   tox
    #   virtualenv
smmap==4.0.0
    # via gitdb
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
snooty-lextudio==1.9.2.dev0
    # via sxm_discord (pyproject.toml)
snowballstemmer==2.1.0
    # via sphinx
soupsieve==2.2.1
    # via beautifulsoup4
sphinx==4.1.1
    # via
    #   furo
    #   sxm_discord (pyproject.toml)
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlalchemy==1.4.21
    # via sxm-player
sqlalchemy-stubs==0.4
    # via sxm_discord (pyproject.toml)
stevedore==3.3.0
    # via bandit
sxm==0.2.5
    # via sxm-player
sxm-player==0.2.1
    # via sxm_discord (pyproject.toml)
tabulate==0.8.9
    # via sxm_discord (pyproject.toml)
tenacity==8.0.1
    # via sxm
termcolor==1.1.0
    # via pytest-sugar
toml==0.10.2
    # via
    #   flit
    #   flit-core
    #   mypy
    #   pep517
    #   pylint
    #   pyproject-flake8
    #   pytest
    #   pytest-cov
    #   snooty-lextudio
    #   tox
tomli==1.0.4
    # via black
tox==3.24.0
    # via sxm_discord (pyproject.toml)
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
typer==0.3.2
    # via
    #   sxm
    #   sxm-player
types-click==7.1.2
    # via sxm_discord (pyproject.toml)
types-psutil==5.8.0
    # via sxm_discord (pyproject.toml)
types-pyyaml==5.4.3
    # via sxm_discord (pyproject.toml)
types-tabulate==0.1.1
    # via sxm_discord (pyproject.toml)
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   discord.py-stubs
    #   mypy
    #   pydantic
    #   snooty-lextudio
    #   sqlalchemy-stubs
    #   sxm_discord (pyproject.toml)
ua-parser==0.10.0
    # via sxm
ujson==1.35
    # via python-jsonrpc-server
urllib3==1.26.6
    # via requests
virtualenv==20.6.0
    # via tox
watchdog==1.0.2
    # via snooty-lextudio
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.36.2
    # via pip-tools
wrapt==1.12.1
    # via astroid
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
